% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmoments.R
\name{logmoments_obj}
\alias{logmoments_obj}
\title{Create a logmoments object}
\usage{
logmoments_obj(x)
}
\arguments{
\item{x}{Numeric vector of positive values}
}
\value{
An object of class "logmoments"
}
\description{
Create a logmoments object
}
\examples{
lc <- logmoments_obj(c(1,2,3,4))
print(lc)
}
