% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmoments.R
\name{print.logmoments}
\alias{print.logmoments}
\title{Print method for logmoments objects}
\usage{
\method{print}{logmoments}(x, ...)
}
\arguments{
\item{x}{An object of class "logmoments"}
}
\description{
Print method for logmoments objects
}
